Code cleaning of KafkaConsumer09.
